application/requirements.txt
Flask
gunicorn
boto3
psycopg2-binary
# python-dotenv # Optional, if you use it for local .env files